refactor: 拆分 Gloden-Layout 扩展脚本
