+++ src/Extensions/Components/BootstrapBlazor.Dock/Components/DockView/DockView.razor.js
@@ -1,6 +1,6 @@
-﻿import "../../js/golden-layout.js"
-import { addLink } from '../../../BootstrapBlazor/modules/utility.js'
+﻿import { addLink } from '../../../BootstrapBlazor/modules/utility.js'
 import Data from '../../../BootstrapBlazor/modules/data.js'
+import Dock from "../../js/golden-layout-extensions.js"
 
 export async function init(id, option, invoke) {
     const el = document.getElementById(id)
@@ -18,8 +18,7 @@ export async function init(id, option, invoke) {
         invoke.invokeMethodAsync(option.visibleChangedCallback, title, visible)
     }
 
-    hackGoldenLayout(dock)
-    const layout = createGoldenLayout(option, el)
+    const layout = Dock.createDock(dock, option);
     dock.layout = layout
     layout.on('initialised', () => {
         saveConfig(option, layout)
@@ -183,14 +182,7 @@ export function dispose(id) {
     }
 
     dock.eventsData.clear()
-    dock.layout.destroy()
-
-    if (goldenLayout.bb_docks !== void 0) {
-        const index = goldenLayout.bb_docks.indexOf(dock);
-        if (index > 0) {
-            goldenLayout.bb_docks.splice(index, 1);
-        }
-    }
+    Dock.dispose(dock)
 }
 
 const lockDock = dock => {
@@ -332,31 +324,6 @@ const getAllItemsByType = (type, parent) => {
     return items
 }
 
-const createGoldenLayout = (option, el) => {
-    const config = getConfig(option)
-
-    if (option.lock) {
-        getAllItemsByType('component', option).forEach(i => {
-            i.componentState.lock = option.lock
-        })
-    }
-
-    const layout = new goldenLayout.GoldenLayout(config, el)
-
-    layout.registerComponentFactoryFunction("component", (container, state) => {
-        const el = document.getElementById(state.id)
-        if (el) {
-            el.classList.remove('d-none')
-            if (state.class) {
-                container.element.classList.add(state.class)
-            }
-            container.element.append(el)
-        }
-    })
-    layout.resizeWithContainerAutomatically = true
-    return layout
-}
-
 const closeItem = (dock, component) => {
     const { template } = dock;
     const item = document.getElementById(component.id)
@@ -366,48 +333,6 @@ const closeItem = (dock, component) => {
     const parent = component.parent
     parent.removeChild(component)
 
-}
-
-const getConfig = option => {
-    option = {
-        enableLocalStorage: false,
-        layoutConfig: null,
-        name: 'default',
-        ...option
-    }
-
-    let config = null
-    let layoutConfig = option.layoutConfig;
-    if (layoutConfig === null && option.enableLocalStorage) {
-        layoutConfig = localStorage.getItem(option.localStorageKey);
-    }
-    if (layoutConfig) {
-        // 当tab全部关闭时，没有root节点
-        const configItem = JSON.parse(layoutConfig)
-        if (configItem.root) {
-            config = configItem
-            resetComponentId(config, option)
-        }
-    }
-
-    return {
-        ...(config || { content: [] }),
-        ...{
-            dimensions: {
-                borderWidth: 5,
-                minItemHeight: 10,
-                minItemWidth: 10,
-                headerHeight: 25
-            },
-            labels: {
-                close: 'close',
-                maximise: 'maximise',
-                minimise: 'minimise',
-                popout: false
-            }
-        },
-        ...option
-    }
 }
 
 const indexOfKey = (key, option) => {
@@ -434,104 +359,6 @@ const removeConfig = option => {
     }
 }
 
-const resetComponentId = (config, option) => {
-    // 本地配置
-    const localComponents = getAllItemsByType('component', config.root)
-    // 服务器端配置
-    const serverItems = getAllItemsByType('component', option)
-    localComponents.forEach(com => {
-        const item = serverItems.find(i => i.componentState.key === com.componentState.key)
-        if (item) {
-            com.id = item.id
-            com.title = item.title
-            com.componentState = item.componentState
-            com.componentState.lock = com.componentState.lock || item.componentState.lock
-        }
-        else {
-            // 本地存储中有，配置中没有，需要显示这个组件，通过 key 来定位新 Component
-            const newEl = document.querySelector(`[data-bb-key='${com.componentState.key}']`) || document.querySelector(`[data-bb-title='${com.componentState.key}']`)
-            if (newEl) {
-                com.id = newEl.getAttribute('id')
-                com.title = newEl.getAttribute('data-bb-title')
-                com.componentState.id = com.id
-
-                option.invokeVisibleChangedCallback(com.title, true)
-            }
-            else {
-                removeContent(config.root.content, com)
-
-                // remove empty stack
-                config.root.content.filter(v => v.content.length === 0).forEach(v => {
-                    const index = config.root.content.indexOf(v)
-                    if (index > -1) {
-                        config.root.content.splice(index, 1)
-                    }
-                })
-            }
-        }
-    })
-
-    serverItems.forEach(item => {
-        // 更新服务器端组件可见状态
-        const com = localComponents.find(i => i.componentState.key === item.componentState.key)
-        option.invokeVisibleChangedCallback(item.title, com !== void 0)
-
-        // 本地存储中没有，配置中有
-        if (com === void 0) {
-            if (config.root.content.length > 0) {
-                if (config.root.type === 'stack' && config.root.content.length === 1 && option.content.length === 1) {
-                    config.root.type = option.content[0].type;
-                    config.root.content = option.content[0].content
-                }
-                else {
-                    config.root.content.push(createItem(item))
-                }
-            }
-            else {
-                config.root.type = option.content[0].type;
-                config.root.content = option.content[0].content;
-            }
-        }
-    })
-
-    // set stack headers
-    // 本地配置
-    const localStacks = getAllItemsByType('stack', config.root)
-    // 服务器端配置
-    const serverStacks = getAllItemsByType('stack', option)
-
-    localStacks.forEach(s => {
-        const stack = {
-            hasHeaders: true,
-            ...findStack(s, serverStacks),
-        };
-        s.header = {
-            ...s.header,
-            show: stack.hasHeaders
-        }
-    });
-}
-
-const createItem = item => ({
-    type: 'component',
-    content: [],
-    title: item.title,
-    id: item.id,
-    componentType: 'component',
-    componentState: item.componentState
-});
-
-const findStack = (stack, stacks) => {
-    let find = null;
-    for (let com of stack.content) {
-        find = stacks.find(s => s.content.find(c => c.componentState.key === com.componentState.key));
-        if (find) {
-            break;
-        }
-    }
-    return find;
-}
-
 const removeContent = (content, item) => {
     content.forEach(v => {
         if (Array.isArray(v.content)) {
@@ -544,56 +371,4 @@ const removeContent = (content, item) => {
             }
         }
     })
-}
-
-const hackGoldenLayout = dock => {
-    if (goldenLayout.bb_docks === void 0) {
-        goldenLayout.bb_docks = [];
-    }
-    goldenLayout.bb_docks.push(dock);
-
-    if (!goldenLayout.isHack) {
-        goldenLayout.isHack = true
-
-        // hack Tab
-        goldenLayout.Tab.prototype.onCloseClick = function () {
-            const component = document.getElementById(this.componentItem.id)
-            const title = this.componentItem.title
-
-            this.notifyClose();
-            this._layoutManager.emit('tabClosed', component, title)
-        }
-
-        // hack RowOrColumn
-        const originSplitterDragStop = goldenLayout.RowOrColumn.prototype.onSplitterDragStop
-        goldenLayout.RowOrColumn.prototype.onSplitterDragStop = function (splitter) {
-            originSplitterDragStop.call(this, splitter)
-            this.layoutManager.emit('splitterDragStop')
-        }
-
-        const originprocessTabDropdownActiveChanged = goldenLayout.Header.prototype.processTabDropdownActiveChanged
-        goldenLayout.Header.prototype.processTabDropdownActiveChanged = function () {
-            originprocessTabDropdownActiveChanged.call(this)
-
-            this._closeButton.onClick = function (ev) {
-                // find own dock
-                const dock = goldenLayout.bb_docks.find(i => i.layout === this._header.layoutManager);
-                const eventsData = dock.eventsData
-
-                const tabs = this._header.tabs.map(tab => {
-                    return { element: tab.componentItem.element, title: tab.componentItem.title }
-                })
-                if (!eventsData.has(this._header.parent)) {
-                    this._pushEvent(ev)
-
-                    const handler = setTimeout(() => {
-                        clearTimeout(handler)
-                        tabs.forEach(tab => {
-                            this._header.layoutManager.emit('tabClosed', tab.element, tab.title)
-                        })
-                    }, 100)
-                }
-            }
-        }
-    }
 }
